refactor(testcontainers): move the JUnit 5 extension to org.jdbi.testing.junit
The v4 rename moved the core testing extensions from
org.jdbi.v3.testing.junit5 to org.jdbi.testing.junit (JUnit 4 support
is gone, so the '5' is redundant), but the Testcontainers extension was
left behind at org.jdbi.testing.junit5.tc. Align it to
org.jdbi.testing.junit.tc so the whole testing surface lives under one
package and the v4 upgrade is a uniform junit5 -> junit move. Pre-release
and riding the existing package-rename churn, so it costs migrators
nothing extra. Also fixes stale testing/junit5/tc javadoc links in the
guide (one sentence already mixed junit/tc and junit5/tc).
