+++ oracle12/src/test/java/org/jdbi/oracle12/TestGetGeneratedKeysOracle.java
@@ -25,7 +25,7 @@ import org.jdbi.sqlobject.statement.GetGeneratedKeys;
 import org.jdbi.sqlobject.statement.SqlQuery;
 import org.jdbi.sqlobject.statement.SqlUpdate;
 import org.jdbi.testing.junit.JdbiExtension;
-import org.jdbi.testing.junit5.tc.JdbiTestcontainersExtension;
+import org.jdbi.testing.junit.tc.JdbiTestcontainersExtension;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
