+++ oracle12/src/test/java/org/jdbi/oracle12/TestOracleInOutParameter.java
@@ -18,7 +18,7 @@ import java.sql.Types;
 import org.jdbi.core.statement.Call;
 import org.jdbi.core.statement.OutParameters;
 import org.jdbi.testing.junit.JdbiExtension;
-import org.jdbi.testing.junit5.tc.JdbiTestcontainersExtension;
+import org.jdbi.testing.junit.tc.JdbiTestcontainersExtension;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
