+++ oracle12/src/test/java/org/jdbi/oracle12/TestOraclePluginIntegration.java
@@ -15,7 +15,7 @@ package org.jdbi.oracle12;
 
 import org.jdbi.core.Handle;
 import org.jdbi.testing.junit.JdbiExtension;
-import org.jdbi.testing.junit5.tc.JdbiTestcontainersExtension;
+import org.jdbi.testing.junit.tc.JdbiTestcontainersExtension;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
