+++ oracle12/src/test/java/org/jdbi/oracle12/TestOracleReturning.java
@@ -21,7 +21,7 @@ import org.jdbi.core.Something;
 import org.jdbi.core.statement.Update;
 import org.jdbi.sqlobject.SqlObjectPlugin;
 import org.jdbi.testing.junit.JdbiExtension;
-import org.jdbi.testing.junit5.tc.JdbiTestcontainersExtension;
+import org.jdbi.testing.junit.tc.JdbiTestcontainersExtension;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
