+++ oracle12/src/test/java/org/jdbi/oracle12/TestOutparameterCursor.java
@@ -21,7 +21,7 @@ import org.jdbi.core.mapper.RowMapper;
 import org.jdbi.core.mapper.reflect.ConstructorMapper;
 import org.jdbi.core.statement.Call;
 import org.jdbi.testing.junit.JdbiExtension;
-import org.jdbi.testing.junit5.tc.JdbiTestcontainersExtension;
+import org.jdbi.testing.junit.tc.JdbiTestcontainersExtension;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
