+++ oracle12/src/test/java/org/jdbi/oracle12/TestScript.java
@@ -25,7 +25,7 @@ import org.jdbi.core.statement.Query;
 import org.jdbi.core.statement.Script;
 import org.jdbi.core.statement.StatementContext;
 import org.jdbi.testing.junit.JdbiExtension;
-import org.jdbi.testing.junit5.tc.JdbiTestcontainersExtension;
+import org.jdbi.testing.junit.tc.JdbiTestcontainersExtension;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
