+++ testcontainers/src/main/java/org/jdbi/testing/junit/tc/EmbeddedUtil.java
@@ -12,7 +12,7 @@
  * limitations under the License.
  */
 
-package org.jdbi.testing.junit5.tc;
+package org.jdbi.testing.junit.tc;
 
 import java.util.concurrent.ThreadLocalRandom;
