+++ testcontainers/src/test/java/org/jdbi/testcontainers/TestMSSQLStoredProcedure.java
@@ -21,7 +21,7 @@ import org.jdbi.core.statement.Call;
 import org.jdbi.core.statement.OutParameters;
 import org.jdbi.sqlobject.SqlObjectPlugin;
 import org.jdbi.testing.junit.JdbiExtension;
-import org.jdbi.testing.junit5.tc.JdbiTestcontainersExtension;
+import org.jdbi.testing.junit.tc.JdbiTestcontainersExtension;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
