+++ testcontainers/src/test/java/org/jdbi/testcontainers/mysql/TestMySQL.java
@@ -23,7 +23,7 @@ import org.jdbi.sqlobject.customizer.Bind;
 import org.jdbi.sqlobject.customizer.Define;
 import org.jdbi.sqlobject.statement.SqlQuery;
 import org.jdbi.testing.junit.JdbiExtension;
-import org.jdbi.testing.junit5.tc.JdbiTestcontainersExtension;
+import org.jdbi.testing.junit.tc.JdbiTestcontainersExtension;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
