+++ testcontainers/src/test/java/org/jdbi/testing/junit/tc/CockroachJdbiTestContainersExtensionTest.java
@@ -11,25 +11,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.testing.junit5.tc;
-
+package org.jdbi.testing.junit.tc;
 
 import org.junit.jupiter.api.Tag;
-import org.junit.jupiter.api.condition.EnabledOnOs;
+import org.testcontainers.containers.CockroachContainer;
 import org.testcontainers.containers.JdbcDatabaseContainer;
-import org.testcontainers.containers.PostgreSQLContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.utility.DockerImageName;
 
 @Tag("slow")
 @Testcontainers
-@EnabledOnOs(architectures = { "x86_64", "amd64" })
-class PostGisJdbiTestContainersExtensionTest extends AbstractJdbiTestcontainersExtensionTest {
+class CockroachJdbiTestContainersExtensionTest extends AbstractJdbiTestcontainersExtensionTest {
 
     @Container
-    static JdbcDatabaseContainer<?> dbContainer = new PostgreSQLContainer<>(
-        DockerImageName.parse("postgis/postgis:13-3.3-alpine").asCompatibleSubstituteFor("postgres"));
+    static JdbcDatabaseContainer<?> dbContainer = new CockroachContainer("cockroachdb/cockroach:latest");
 
     @Override
     JdbcDatabaseContainer<?> getDbContainer() {
