+++ testcontainers/src/test/java/org/jdbi/testing/junit/tc/DB2JdbiTestContainersExtensionTest.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.testing.junit5.tc;
+package org.jdbi.testing.junit.tc;
 
 import java.nio.charset.Charset;
