+++ testcontainers/src/test/java/org/jdbi/testing/junit/tc/OracleXeJdbiTestContainersExtensionTest.java
@@ -11,22 +11,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.testing.junit5.tc;
-
+package org.jdbi.testing.junit.tc;
 
 import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.condition.EnabledOnOs;
 import org.testcontainers.containers.JdbcDatabaseContainer;
+import org.testcontainers.containers.OracleContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.tidb.TiDBContainer;
 
 @Tag("slow")
 @Testcontainers
-class TiDBJdbiTestContainersExtensionTest extends AbstractJdbiTestcontainersExtensionTest {
+@EnabledOnOs(architectures = {"x86_64", "amd64"})
+class OracleXeJdbiTestContainersExtensionTest extends AbstractJdbiTestcontainersExtensionTest {
 
     @Container
-    static JdbcDatabaseContainer<?> dbContainer = new TiDBContainer("pingcap/tidb");
-
+    static JdbcDatabaseContainer<?> dbContainer = new OracleContainer("gvenzl/oracle-xe:slim-faststart");
 
     @Override
     JdbcDatabaseContainer<?> getDbContainer() {
