+++ testcontainers/src/test/java/org/jdbi/testing/junit/tc/TrinoJdbiTestContainersExtensionTest.java
@@ -11,24 +11,31 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.testing.junit5.tc;
+package org.jdbi.testing.junit.tc;
 
+
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Tag;
 import org.testcontainers.containers.JdbcDatabaseContainer;
-import org.testcontainers.containers.MSSQLServerContainer;
+import org.testcontainers.containers.TrinoContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
 @Tag("slow")
 @Testcontainers
-class MSSQLJdbiTestContainersExtensionTest extends AbstractJdbiTestcontainersExtensionTest {
+class TrinoJdbiTestContainersExtensionTest extends AbstractJdbiTestcontainersExtensionTest {
 
     @Container
-    static JdbcDatabaseContainer<?> dbContainer = new MSSQLServerContainer<>("mcr.microsoft.com/mssql/server:2022-latest")
-        .acceptLicense();
+    static JdbcDatabaseContainer<?> dbContainer = new TrinoContainer("trinodb/trino");
 
     @Override
     JdbcDatabaseContainer<?> getDbContainer() {
         return dbContainer;
+    }
+
+    @BeforeEach
+    public void beforeEach() throws Exception {
+        // we still love you, trino. But you are a mess.
+        Thread.sleep(5_000);
     }
 }
